refactor: made hooks in button main and hook for localStorage
