+++ .husky/pre-push
@@ -1,0 +1,2 @@
+npx lint-staged
+npm run lint
