+++ src/App.css
@@ -24,28 +24,6 @@ body {
   color: #393ff4;
 }
 
-.main {
-  border: 1px solid skyblue;
-  border-radius: 5px;
-  padding: 5px 10px;
-  min-height: 80vh;
-  p {
-    color: rgb(253, 224, 0);
-  }
-  div {
-    margin: 2px 0;
-    line-height: 0;
-    border: 1px solid #00ffc4;
-  }
-  h3 {
-    color: #00ffc4;
-  }
-}
-
-.title-main {
-  color: #e69d0c;
-}
-
 .searchBTN {
   margin-left: 5px;
   background-color: #61dafbaa;
