+++ src/components/Main/main.module.css
@@ -1,0 +1,21 @@
+.main {
+  border: 1px solid skyblue;
+  border-radius: 5px;
+  padding: 5px 10px;
+  min-height: 80vh;
+  p {
+    color: rgb(253, 224, 0);
+  }
+  div {
+    margin: 2px 0;
+    line-height: 0;
+    border: 1px solid #00ffc4;
+  }
+  h3 {
+    color: #00ffc4;
+  }
+}
+
+.title-main {
+  color: #e69d0c;
+}
